add IndivisualPoolStats
